requirements.txt
kivy[base]
schema==0.7.5
pytest==7.2.2
pytest-asyncio==0.21.0
requests==2.28.2
pytest-mock==3.10.0
PyAutoGUI==0.9.53
pyyaml==6.0
